requirements.txt
mljar-supervised
fastapi
uvicorn
scikit-learn
joblib
